# feature(remap source): Emitting multiple events