Few remaining TODOs, but this is almost there. The approach I took was to introduce an "adapter" for `vector::Event` rather than having `event::Metric` and `event::LogEvent` implement `vrl::Target` directly. This allowed me to decouple the implementation for `LogEvent` from the underlying data structure; instead taking the shortcut of just representing it as a `vrl::Value`, which already implements `vrl::Target`, and then marshaling back into a `Vec<Event>` to map arrays.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

<!--
**Your PR title must conform to the conventional commit spec!**

  <type>!?(<scope>): <description>

  * `type` = chore, enhancement, feat, fix
  * `!` = signals a breaking change
  * `scope` = https://github.com/timberio/vector/blob/master/.github/semantic.yml#L4
  * `description` = short description of the change

Examples:

  * enhancement(file source): Added `sort` option to sort discovered files
  * feat(new source): Initial `statsd` source
  * fix(file source): Fixed a bug discovering new files
  * chore(external docs): Clarified `batch_size` option
-->