@StephenWakely 2021-04-29
This is sweet! 

It will be very interesting to benchmark this as per https://github.com/timberio/vector/pull/5374#issuecomment-816314169. I suspect it will do wonders for a Vrl heavy config since from what I can gather we are no longer paying the cost of converting between a Vrl Value and a Vector Value with each call.

The multiple events is just an added bonus!