@jszwedko 2021-04-29
I ran a few benchmarks and found some interesting results:

* It seems like inserting, via the `Target` implementation, is generally faster for `event::LogEvent` than it is `vrl::Value`. The more `insert`s I do, the bigger the gap between master and this branch.  I'm curious if you have thoughts there @StephenWakely . There could be something else I'm doing here that is causing that behavior though.

Benchmarks using a config that looks like:

```toml
[sources.in]
type = "stdin"

[transforms.remap]
type = "remap"
inputs = ["in"]
drop_on_error = true
source = """
.foo0 = "bar"
# adding additional inserts here
"""

[sinks.out]
type = "blackhole"
inputs = ["remap"]
```

Master, empty remap transform:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ./target/release/vector-master-0240ae601 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      25976.082806      task-clock:u (msec)       #    2.533 CPUs utilized            ( +-  0.21% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              5388      page-faults:u             #    0.207 K/sec                    ( +-  2.80% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      10.256256539 seconds time elapsed                                          ( +-  0.21% )
```

This branch, empty remap transform:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ./target/release/vector-emit-multiple-3ca72b111 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      30524.937509      task-clock:u (msec)       #    2.744 CPUs utilized            ( +-  0.57% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              4951      page-faults:u             #    0.162 K/sec                    ( +-  0.33% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      11.123577205 seconds time elapsed                                          ( +-  0.55% )
```

A roughly 8% penalty.

Master, one field assignment:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ./target/release/vector-master-0240ae601 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      27542.389545      task-clock:u (msec)       #    2.626 CPUs utilized            ( +-  0.55% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              5251      page-faults:u             #    0.191 K/sec                    ( +-  2.67% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      10.488638371 seconds time elapsed                                          ( +-  0.78% )
```

This branch, one field assignment:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ./target/release/vector-emit-multiple-3ca72b111 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      33143.892345      task-clock:u (msec)       #    2.846 CPUs utilized            ( +-  0.38% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              5122      page-faults:u             #    0.155 K/sec                    ( +-  1.23% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      11.646199998 seconds time elapsed                                          ( +-  0.48% )
```

Master, 10 field assignments:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ./target/release/vector-master-0240ae601 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      36705.275313      task-clock:u (msec)       #    2.968 CPUs utilized            ( +-  0.37% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              5119      page-faults:u             #    0.139 K/sec                    ( +-  0.77% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      12.365062207 seconds time elapsed                                          ( +-  0.39% )
```

This branch, 10 field assignments:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ./target/release/vector-emit-multiple-3ca72b111 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      49701.886379      task-clock:u (msec)       #    2.876 CPUs utilized            ( +-  0.21% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              4174      page-faults:u             #    0.084 K/sec                    ( +-  0.16% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      17.278687059 seconds time elapsed                                          ( +-  0.33% )
```

One thought I had for not paying that penalty for the empty transform is to delay converting `LogEvent` into a `vrl::Value` until the user actually assigns `.` when the conversion between `vrl::Value` and `event::Value`  would happen anyway.