@StephenWakely 2021-04-29
> I ran a few benchmarks and found some interesting results:

That is quite the opposite to what I was expecting!

The Vrl Value insert and the Vector Value insert are doing things differently. My first suspicion is that Vector is doing something much more efficient. It will be interesting to compare the two to see what the differences are.

It is a shame that we have two different implementations of exactly the same thing. At some point it would be nice to reconcile the two..