@jszwedko 2021-05-03
I think this is ready for review now. I updated the implementation and it seems to suffer less overhead (on the order of ~5%; independent of the number fields inserted).

Example:

```toml
[sources.in]
type = "stdin"

[transforms.remap]
type = "remap"
inputs = ["in"]
source = """
[sources.in]
type = "stdin"

[transforms.remap]
type = "remap"
inputs = ["in"]
source = """
. |= object!(parse_json!(.message))
. = .events
"""

[sinks.out]
type = "console"
encoding.codec = "json"
inputs = ["remap"]"""

[sinks.out]
type = "console"
encoding.codec = "json"
inputs = ["remap"]
```

```
➜  vector git:(emit-multiple) ✗ echo '{ "events": [1, "2", true, null, { "message": "hello world"} ]}' | ./target/debug/vector --config /tmp/remap.toml
May 03 20:53:33.057  INFO vector::app: Log level is enabled. level="vector=info,codec=info,vrl=info,file_source=info,tower_limit=trace,rdkafka=info"
May 03 20:53:33.057  INFO vector::app: Loading configs. path=[("/tmp/remap.toml", None)]
May 03 20:53:33.088  INFO vector::sources::stdin: Capturing STDIN.
May 03 20:53:33.088  INFO vector::topology: Running healthchecks.
May 03 20:53:33.089  INFO vector::topology::builder: Healthcheck: Passed.
May 03 20:53:33.089  INFO vector::topology: Starting source. name="in"
May 03 20:53:33.089  INFO vector::topology: Starting transform. name="remap"
May 03 20:53:33.089  INFO vector::topology: Starting sink. name="out"
May 03 20:53:33.090  INFO vector: Vector has started. version="0.14.0" git_version="v0.12.0-289-g90ad54a" released="Mon, 03 May 2021 19:23:17 +0000" arch="x86_64"
May 03 20:53:33.091  INFO vector::shutdown: All sources have finished.
May 03 20:53:33.091  INFO source{component_kind="source" component_name=in component_type=stdin}: vector::sources::stdin: Finished sending.
May 03 20:53:33.091  INFO vector: Vector has stopped.
May 03 20:53:33.093  INFO vector::topology: Shutting down... Waiting on running components. remaining_components="out, remap" time_remaining="59 seconds left"
{"message":1}
{"message":"2"}
{"message":true}
{"message":null}
{"message":"hello world"}
```

Note here I took the suggestion made in #7038 to just use the value as the `message` key for non-object types.

Benchmarks (see above for config):

```
Master empty remap transform:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ~/bin/vector-master-0240ae601 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      26262.471163      task-clock:u (msec)       #    2.530 CPUs utilized            ( +-  0.59% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              5474      page-faults:u             #    0.208 K/sec                    ( +-  2.26% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      10.381405504 seconds time elapsed                                          ( +-  0.58% )
```

This branch:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ./target/release/vector-emit-multiple-effc6b3b2 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      29890.939533      task-clock:u (msec)       #    2.723 CPUs utilized            ( +-  0.27% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              5364      page-faults:u             #    0.179 K/sec                    ( +-  2.20% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      10.975345426 seconds time elapsed                                          ( +-  0.27% )
```

Master 10 field inserts:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ~/bin/vector-master-0240ae601 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      37098.020793      task-clock:u (msec)       #    2.940 CPUs utilized            ( +-  0.33% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              5119      page-faults:u             #    0.138 K/sec                    ( +-  1.25% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      12.618747150 seconds time elapsed                                          ( +-  0.36% )
```

This branch:

```
 Performance counter stats for 'bash -c yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 | ./target/release/vector-emit-multiple-effc6b3b2 --config ~/tmp/remap.toml > /dev/null' (10 runs):

      38343.198227      task-clock:u (msec)       #    2.920 CPUs utilized            ( +-  2.58% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
              5210      page-faults:u             #    0.136 K/sec                    ( +-  1.44% )
   <not supported>      cycles:u
   <not supported>      instructions:u
   <not supported>      branches:u
   <not supported>      branch-misses:u

      13.129863929 seconds time elapsed                                          ( +-  2.59% )
```

I expected a little bit of overhead just creating additional data structures like the iterators, but 5% is more than I expected.

I ran `perf diff` after running `sudo perf record -F 999 -a -g --  ./target/release/vector--config ~/tmp/remap.toml < /tmp/input.txt` on master vs this branch and saw:

```
               +2.53%  vector-emit-multiple      [.] <twox_hash::sixty_four::XxHash64 as core::hash::Hasher>::write
    68.89%     -1.68%  [kernel.kallsyms]         [k] native_safe_halt
               +1.30%  vector-emit-multiple      [.] tracing_subscriber::registry::sharded::Registry::get
               +0.77%  vector-emit-multiple      [.] metrics_tracing_context::TracingContext<R,F>::enhance_key
               +0.71%  vector-emit-multiple      [.] metrics_tracing_context::tracing_integration::MetricsLayer<S>::with_labels
               +0.70%  vector-emit-multiple      [.] dashmap::DashMap<K,V,S>::entry
               +0.58%  vector-emit-multiple      [.] serde_json::ser::format_escaped_str_contents
               +0.53%  vector-emit-multiple      [.] <tracing_subscriber::layer::Layered<L,S> as tracing_core::subscriber::Subscriber>::try_close
               +0.52%  vector-emit-multiple      [.] <tracing_subscriber::registry::sharded::Registry as tracing_core::subscriber::Subscriber>::current_span
               +0.49%  vector-emit-multiple      [.] metrics_tracing_context::TracingContext$LT$R$C$F$GT$::enhance_labels::_$u7b$$u7b$closure$u7d$$u7d$::hf1170eadf40f812f
               +0.49%  vector-emit-multiple      [.] <twox_hash::sixty_four::XxHash64 as core::hash::Hasher>::finish
               +0.48%  vector-emit-multiple      [.] <tracing_subscriber::registry::sharded::Registry as tracing_core::subscriber::Subscriber>::clone_span
               +0.41%  vector-emit-multiple      [.] <metrics::key::KeyData as core::hash::Hash>::hash
               +0.38%  vector-emit-multiple      [.] std::sys_common::backtrace::__rust_begin_short_backtrace
               +0.36%  vector-emit-multiple      [.] futures_channel::mpsc::Receiver<T>::next_message
               +0.35%  vector-emit-multiple      [.] <vector::metrics::recorder::VectorRecorder as metrics::recorder::Recorder>::increment_counter
               +0.23%  vector-emit-multiple      [.] <core::str::lossy::Utf8LossyChunksIter as core::iter::traits::iterator::Iterator>::next
               +0.22%  vector-emit-multiple      [.] vector::event::util::log::insert::map_insert
               +0.22%  vector-emit-multiple      [.] futures_channel::mpsc::Sender<T>::start_send
               +0.21%  vector-emit-multiple      [.] <metrics_tracing_context::TracingContext<R,F> as metrics::recorder::Recorder>::increment_counter
               +0.21%  vector-emit-multiple      [.] <alloc::collections::btree::map::BTreeMap<K,V> as core::clone::Clone>::clone::clone_subtree
               +0.21%  vector-emit-multiple      [.] <vector::event::util::log::path_iter::PathIter as core::iter::traits::iterator::Iterator>::next
     2.57%     +0.20%  [kernel.kallsyms]         [k] __lock_text_start
     1.02%     +0.19%  libc-2.26.so              [.] _int_malloc
     0.61%     +0.19%  libc-2.26.so              [.] cfree@GLIBC_2.2.5
               +0.18%  vector-emit-multiple      [.] regex::re_unicode::Regex::is_match
               +0.16%  vector-emit-multiple      [.] <tracing_subscriber::registry::sharded::Data as tracing_subscriber::registry::SpanData>::extensions
               +0.16%  vector-emit-multiple      [.] chrono::format::format_inner
     1.12%     +0.15%  libc-2.26.so              [.] _int_free
               +0.15%  vector-emit-multiple      [.] <metrics::key::Key as core::cmp::PartialEq>::eq
               +0.13%  vector-emit-multiple      [.] vector::event::value::<impl core::convert::From<vector::event::value::Value> for vrl_compiler::value::Value>::from
               +0.12%  vector-emit-multiple      [.] <core::future::from_generator::GenFuture<T> as core::future::future::Future>::poll
               +0.12%  vector-emit-multiple      [.] <vector::event::log_event::LogEvent as serde::ser::Serialize>::serialize
               +0.11%  vector-emit-multiple      [.] core::fmt::write
               +0.11%  vector-emit-multiple      [.] <metrics::key::NameParts as core::clone::Clone>::clone
               +0.10%  vector-emit-multiple      [.] core::ptr::drop_in_place<metrics::key::Key>
               +0.09%  vector-emit-multiple      [.] <tracing_subscriber::registry::sharded::CloseGuard as core::ops::drop::Drop>::drop
               +0.09%  vector-emit-multiple      [.] core::ptr::drop_in_place<vrl_compiler::value::Value>
               +0.09%  vector-emit-multiple      [.] alloc::collections::btree::map::BTreeMap<K,V>::insert
               +0.08%  vector-emit-multiple      [.] bytes::bytes::promotable_even_drop
               +0.08%  vector-emit-multiple      [.] alloc::raw_vec::finish_grow
```

`input.txt` was generated via `yes "{\"appname\":\"su\",\"hostname\":\"initech.io\",\"message\":\"TPS report missing cover sheet\",\"msgid\":\"ID81\",\"procid\":4015,\"timestamp\":\"2020-12-19T21:48:09.004Z\",\"version\":3}" | head -n 1000000 > /tmp/input.txt`.

The biggest percent differences actually seemed to be at the metric layer (`twox_hash` is used at that layer) which surprised me. I'm open to ideas or thoughts here, but it seems reasonable to me to proceed.